# Don't use serde_json to read the list of named entities for codegen