--- a/markup5ever/Cargo.toml
+++ b/markup5ever/Cargo.toml
@@ -22,6 +22,3 @@ log = "0.4"
 [build-dependencies]
 string_cache_codegen = "0.5.1"
 phf_codegen = "0.8"
-serde = "1.0"
-serde_derive = "1.0"
-serde_json = "1.0"