--- a/markup5ever/build.rs
+++ b/markup5ever/build.rs
@@ -9,15 +9,14 @@
 
 extern crate phf_codegen;
 extern crate string_cache_codegen;
-#[macro_use]
-extern crate serde_derive;
-extern crate serde_json;
 
 use std::collections::HashMap;
 use std::env;
 use std::fs::File;
 use std::io::{BufRead, BufReader, BufWriter, Write};
 use std::path::Path;
+
+mod entities;
 
 static NAMESPACES: &'static [(&'static str, &'static str)] = &[
     ("", ""),
@@ -34,10 +33,7 @@ fn main() {
     let generated = Path::new(&env::var("OUT_DIR").unwrap()).join("generated.rs");
     let mut generated = BufWriter::new(File::create(&generated).unwrap());
 
-    let manifest_dir = env::var("CARGO_MANIFEST_DIR").unwrap();
-
     named_entities_to_phf(
-        &Path::new(&manifest_dir).join("data").join("entities.json"),
         &Path::new(&env::var("OUT_DIR").unwrap()).join("named_entities.rs"),
     );
 
@@ -89,27 +85,14 @@ fn main() {
     writeln!(generated, "}}").unwrap();
 }
 
-fn named_entities_to_phf(from: &Path, to: &Path) {
-    // A struct matching the entries in entities.json.
-    #[derive(Deserialize, Debug)]
-    struct CharRef {
-        codepoints: Vec<u32>,
-        //characters: String,  // Present in the file but we don't need it
-    }
-
-    let entities: HashMap<String, CharRef> =
-        serde_json::from_reader(&mut File::open(from).unwrap()).unwrap();
-    let mut entities: HashMap<&str, (u32, u32)> = entities
+fn named_entities_to_phf(to: &Path) {
+    let mut entities: HashMap<&str, (u32, u32)> = entities::NAMED_ENTITIES
         .iter()
-        .map(|(name, char_ref)| {
+        .map(|(name, cp1, cp2)| {
             assert!(name.starts_with("&"));
-            assert!(char_ref.codepoints.len() <= 2);
             (
                 &name[1..],
-                (
-                    char_ref.codepoints[0],
-                    *char_ref.codepoints.get(1).unwrap_or(&0),
-                ),
+                (*cp1, *cp2),
             )
         })
         .collect();