This can have the effect of taking out serde_derive from the critical
path of crates that only use markup5ever.  In the case of librsvg,
this reduces build time in release mode by 37 seconds.

Here are two runs of `cargo build -Z timings --release` for librsvg.  Apologies for not attaching the HTML output of `-Z timings`, but github doesn't let one attach HTML files:

* [Unpatched](https://people.gnome.org/~federico/misc/cargo-timing-20191030T151148Z.html) - 149.6s total; markup5ever starts building at around 128s.
* [Patched](https://people.gnome.org/~federico/misc/cargo-timing-20191030T194418Z.html) - 112.5s total; markup5ever starts building at around 51s.