@bors-servo 2019-10-30
:sunny: Test successful - [checks-travis](https://travis-ci.com/servo/html5ever/builds/134277248)
Approved by: jdm
Pushing 9066e3b243de936e8b6277e609fd9cfa9a2d734a to master...